+++ src/community/static/css/Notice.module.css
@@ -36,10 +36,11 @@
 .pageListBox{
     display: flex;
     justify-content: center;
+    padding : 1vh;
 }
 
 .pageButton {
-    font-size: 20px;
+    font-size: 18px;
     padding : 5px;
     color : #BBBBBB;
 }
